+++ src/delimited_index_connection.cc
@@ -81,8 +81,11 @@ delimited_index_connection::delimited_index_connection(
     return;
   }
 
+  bool has_quoted_newlines = quote != '\0';
+
   // Parse header
-  size_t start = find_first_line(buf[i], skip_, comment_, skip_empty_rows);
+  size_t start = find_first_line(
+      buf[i], skip_, comment_, skip_empty_rows, has_quoted_newlines);
 
   if (delim == nullptr) {
     delim_ = std::string(1, guess_delim(buf[i], start, 5, sz));
@@ -93,7 +96,7 @@ delimited_index_connection::delimited_index_connection(
   delim_len_ = delim_.length();
 
   size_t first_nl = find_next_newline(
-      buf[i], start, comment, skip_empty_rows, /* embededd_nl */ true);
+      buf[i], start, comment, skip_empty_rows, has_quoted_newlines);
 
   bool single_line = first_nl == buf[i].size() - 1;